ramips: mt7621: Add DNA Valokuitu Plus EX400
Hardware
--------

Specifications:
- Device: DNA Valokuitu Plus EX400
- SoC: MT7621A
- Flash: 256MB NAND
- RAM:  256MB
- Ethernet: Built-in, 2 x 1GbE
- Wifi: MT7603 2.4 GHz, MT7615 5 GHz (4x internal antennas)
- USB: 1x 3.0
- LED: 1x green/red, 1x green
- Buttons: Reset

MAC addresses:

LAN:     u-boot 'ethaddr' (label)
WAN:     label + 1
2.4 GHz: label + 6
5 GHz:   label + 7

Serial:
There is a black block connector next to the red ethernet connector. It
is accessible also through a hole in the casing.

Pinout (TTL 3.3V)
+---+---+
|Tx |Rx |
+---+---+
|Vcc|Gnd|
+---+---+

Firmware
--------

The vendor firmware a fork of OpenWrt (Reboot)  with kernel version
4.4.93. The original firmware has the NAND space split in two (apart from
the first 1MB, which is for uboot) and it uses an u-boot based two-slot
mechanism with a boot failure counter. If booting from one slot fails five
times it will try the other slot. For firmware upgrades the new image is
installed to the slot not being currently used. That mechanism is preserved
for robustness and for easier roll-back to the vendor firmware if need be.

U-boot:
Booting can be halted to U-boot with proper serial access. Tftp and flash
writes are available, but only the first one has been tested.

Installing OpenWrt:
Use the factory image in firmware upgrade page of the vendor web ui.
Alternatively, ssh to the device and use sysupgrade -n from cli.

Going back to manufacturer firmware (example):
 - Get ubifs offset and size from the combined vendor image file
   head -c 1K EX400-X-DNA-4.3.6.100-R-210518_0935.y3 | \
     tr -cd '\11\12\15\40-\176' | grep ubifs
   ubifsofs 1024
   ubifs 29077504

 - Splice out the ubifs image
   dd if=./EX400-X-DNA-4.3.6.100-R-210518_0935.y3 bs=1 skip=1024 \
     count=29077504 of=./image.ubi

 - Install the image (image.ubi) using sysupgrade -n -F or the Luci
   interface
 - Reboot the device and install the vendor image once again image using
   their web ui.

Known issues:
 - MACs for wifi are stored in currently unknown place but it's obtained
   allright
 - kernel and dtb must be placed under /boot/ of rootfs for u-boot to find.
   Currently it does that for all targets of the platform.
 - Overlay file system ends up in looping symbolic links for the upper and
   work directories. This is problematic for clearing the settings or for
   restoring configuration.

Signed-off-by: Mauri Sandberg <maukka@ext.kapsi.fi>