base-files: upgrade: nand: create rootfs_data for ubifs too
In case of ubifs, rootfs volume seems to hog all available space and no
'rootfs_data' volume is being created. This is problematic because
mount_root looks for 'rootfs_data' for writable overlay. When it does not
find it the / is remounted rw without an overlayfs at all.

Creating the rootfs_data for ubifs rootfs fixes the issue.

Signed-off-by: Mauri Sandberg <maukka@ext.kapsi.fi>